Refactor: Centralize duplicate utility functions to utils.js
