+++ docs/javascript/filters.js
@@ -2,7 +2,7 @@
  * @fileoverview Logic for hero listing layout rendering, table sort columns, sidebar drawers, active filter chips, search queries, and recency calculations.
  * @module filters
  */
-import { isHeroOwned } from './utils.js';
+import { isHeroOwned, parseDateString, getDaysAgoClean, getRecencyDot } from './utils.js';
 import { renderGamesList } from './admin.js';
 import { updateRollSettingsBadge } from './randomizer.js';
 export function toggleStagedGamesWinnerOnly(checked) { stateStore.set("stagedGamesWinnerOnly", checked); renderDrawerBody(); }
@@ -458,75 +458,4 @@ export function matchesSearchTerm(c, term) {
         (c.name || "").toLowerCase().includes(clean) ||
         (c.group || "").toLowerCase().includes(clean)
     );
-}
-export function parseDateString(dateString) {
-    if (!dateString) return null;
-    try {
-        let cleanDate = dateString.trim();
-        if (cleanDate && !cleanDate.includes("T"))
-            cleanDate = cleanDate.replace(" ", "T");
-        if (
-            cleanDate &&
-            cleanDate.includes(":") &&
-            !cleanDate.includes("Z") &&
-            !cleanDate.includes("+")
-        )
-            cleanDate += "Z";
-        const dateObj = new Date(cleanDate);
-        return isNaN(dateObj.getTime()) ? null : dateObj;
-    } catch (e) {
-        return null;
-    }
-}
-export function getDaysAgoClean(dateString) {
-    if (!dateString || dateString === "Never") return "";
-    if (dateString === "Unknown") return "Date unknown (historical)";
-    const lastDate = parseDateString(dateString);
-    if (!lastDate) return "";
-    try {
-        const today = new Date();
-        lastDate.setHours(0, 0, 0, 0);
-        today.setHours(0, 0, 0, 0);
-        const diffTime = today - lastDate;
-        const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-        if (diffDays < 0) return "";
-        if (diffDays === 0) return "today";
-        if (diffDays === 1) return "yesterday";
-        return `${diffDays} days ago`;
-    } catch (e) {
-        return "";
-    }
-}
-export function getRecencyDot(lastPlayed) {
-    let recencyDot = "⚫";
-    if (lastPlayed && lastPlayed === "Unknown") {
-        recencyDot = "🔴";
-    } else if (
-        lastPlayed &&
-        lastPlayed !== "Never" &&
-        lastPlayed !== "Unknown"
-    ) {
-        const lastDate = parseDateString(lastPlayed);
-        if (lastDate) {
-            try {
-                const today = new Date();
-                lastDate.setHours(0, 0, 0, 0);
-                today.setHours(0, 0, 0, 0);
-                const diffTime = today - lastDate;
-                const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-                if (diffDays <= 15) {
-                    recencyDot = "🟢";
-                } else if (diffDays <= 60) {
-                    recencyDot = "🟡";
-                } else {
-                    recencyDot = "🔴";
-                }
-            } catch (e) {
-                recencyDot = "⚪";
-            }
-        } else {
-            recencyDot = "⚪";
-        }
-    }
-    return recencyDot;
 }
