+++ docs/javascript/views/rollView.js
@@ -10,7 +10,10 @@ import {
     getHeroProbabilityText, 
     isHeroOwned, 
     getSoftWeight,
-    escapeHtml 
+    escapeHtml,
+    parseDateString,
+    getDaysAgoClean,
+    getRecencyDot
 } from '../utils.js';
 import { updateSegmentedHighlights } from './filterView.js';
 
@@ -252,88 +255,7 @@ export function updatePlayerCardUI(pIdx, finalHero) {
     }
 }
 
-/**
- * Helper to get date parsing (similar to filters.js)
- */
-function parseDateString(dateString) {
-    if (!dateString) return null;
-    try {
-        let cleanDate = dateString.trim();
-        if (cleanDate && !cleanDate.includes("T"))
-            cleanDate = cleanDate.replace(" ", "T");
-        if (
-            cleanDate &&
-            cleanDate.includes(":") &&
-            !cleanDate.includes("Z") &&
-            !cleanDate.includes("+")
-        )
-            cleanDate += "Z";
-        const dateObj = new Date(cleanDate);
-        return isNaN(dateObj.getTime()) ? null : dateObj;
-    } catch (e) {
-        return null;
-    }
-}
 
-/**
- * Decoupled relative elapsed days generator.
- */
-function getDaysAgoClean(dateString) {
-    if (!dateString || dateString === "Never") return "";
-    if (dateString === "Unknown") return "Date unknown (historical)";
-    const lastDate = parseDateString(dateString);
-    if (!lastDate) return "";
-    try {
-        const today = new Date();
-        lastDate.setHours(0, 0, 0, 0);
-        today.setHours(0, 0, 0, 0);
-        const diffTime = today - lastDate;
-        const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-        if (diffDays < 0) return "";
-        if (diffDays === 0) return "today";
-        if (diffDays === 1) return "yesterday";
-        return `${diffDays} days ago`;
-    } catch (e) {
-        return "";
-    }
-}
-
-/**
- * Decoupled recency emoji decorator.
- */
-function getRecencyDot(lastPlayed) {
-    let recencyDot = "⚫";
-    if (lastPlayed && lastPlayed === "Unknown") {
-        recencyDot = "🔴";
-    } else if (
-        lastPlayed &&
-        lastPlayed !== "Never" &&
-        lastPlayed !== "Unknown"
-    ) {
-        const lastDate = parseDateString(lastPlayed);
-        if (lastDate) {
-            try {
-                const today = new Date();
-                lastDate.setHours(0, 0, 0, 0);
-                today.setHours(0, 0, 0, 0);
-                const diffTime = today - lastDate;
-                const diffDays = Math.floor(diffTime / (1000 * 60 * 60 * 24));
-                if (diffDays <= 15) {
-                    recencyDot = "🟢";
-                } else if (diffDays <= 60) {
-                    recencyDot = "🟡";
-                } else {
-                    recencyDot = "🔴";
-                }
-            } catch (e) {
-                recencyDot = "⚪";
-            }
-        } else {
-            recencyDot = "⚪";
-        }
-    }
-    return recencyDot;
-}
 
 /**
  * Collapses the draft row view to show the final selected character's stats and edit triggers.
